+++ thirdparty/bob/start.php
@@ -1,0 +1,6 @@
+<?php
+
+Autoloader::directories(array(
+	__DIR__.DS.'classes',
+	__DIR__.DS.'classes'.DS.'generators'
+));