+++ changes.md
@@ -6,3 +6,7 @@ in some of Laravel's Form methods, this allows us to nest them more easily
 #### Authority
 - Changed config location
 - Changed table name (users => accounts)
+
+#### Bob
+- Changed Autoloader in start.php
+- Chaged template path in Common